// src/app/shared/models/persons.ts
export interface Iperson {
    fname: string,
    lname: string,
    email: string,
    contact: number
}

export interface Student {
    name: string;
    age: number;
    grade: string;
    subject: string;
  }
  
  export interface Book {
    title: string;
    author: string;
    pages: number;
    genre: string;
  }
  
  export interface Employee {
    name: string;
    position: string;
    salary: number;
    department: string;
  }
  
  export interface Product {
    name: string;
    brand: string;
    price: number;
    category: string;
  }
  